# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Сервис для управления ПВЗ и приемкой товаров",
        "title": "Pickup Point API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/dummyLogin": {
            "post": {
                "description": "Получение тестового токена авторизации по роли",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Dummy login",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.dummyLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает JWT токен для аутентификации",
                        "schema": {
                            "$ref": "#/definitions/v1.dummyLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "description": "Аутентификация пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Учетные данные",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает JWT токен для аутентификации",
                        "schema": {
                            "$ref": "#/definitions/v1.loginResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/products": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Добавляет товар в последнюю незакрытую приёмку в указанном ПВЗ. Доступно только для сотрудников ПВЗ. Требуется незакрытая приёмка.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Добавление товара в приёмку",
                "parameters": [
                    {
                        "description": "Данные для добавления товара",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.createProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v1.createProductResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный идентификатор ПВЗ, тип товара или отсутствие открытой приёмки",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/pvz/{pvzId}/delete_last_product": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Удаляет последний добавленный товар в последней незакрытой приёмке указанного ПВЗ. Доступно только для сотрудников ПВЗ. Требуется наличие незакрытой приёмки и хотя бы одного товара в ней.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Удаление последнего добавленного товара",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор ПВЗ (uuid)",
                        "name": "pvzId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном удалении",
                        "schema": {
                            "$ref": "#/definitions/v1.deleteProductResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный идентификатор ПВЗ, отсутствие открытой приёмки или отсутствие товаров в приёмке",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/register": {
            "post": {
                "description": "Регистрация нового пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Возвращает данные зарегистрированного пользователя",
                        "schema": {
                            "$ref": "#/definitions/v1.registerResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Пользователь с таким email уже существует",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httpresponse.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "v1.createProductRequest": {
            "description": "Запрос для добавления товара",
            "type": "object",
            "properties": {
                "pvzId": {
                    "description": "Идентификатор ПВЗ\nformat: uuid",
                    "type": "string"
                },
                "type": {
                    "description": "Тип товара\nenum: электроника, одежда, обувь",
                    "type": "string"
                }
            }
        },
        "v1.createProductResponse": {
            "description": "Ответ с данными о добавленном товаре",
            "type": "object",
            "properties": {
                "dateTime": {
                    "description": "Дата и время добавления товара\nformat: date-time",
                    "type": "string"
                },
                "id": {
                    "description": "Уникальный идентификатор товара\nformat: uuid",
                    "type": "string"
                },
                "receptionId": {
                    "description": "Идентификатор приёмки\nformat: uuid",
                    "type": "string"
                },
                "type": {
                    "description": "Тип товара",
                    "type": "string"
                }
            }
        },
...
        "v1.deleteProductResponse": {
            "description": "Ответ с сообщением об удалении товара",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Сообщение об успешном удалении товара",
                    "type": "string"
                }
            }
        },
        "v1.dummyLoginRequest": {
            "description": "Запрос для получения тестового токена авторизации",
            "type": "object",
            "properties": {
                "role": {
                    "description": "Роль пользователя (employee или moderator)\nenum: employee,moderator",
                    "type": "string"
                }
            }
        },
        "v1.dummyLoginResponse": {
            "description": "Ответ с тестовым токеном авторизации",
            "type": "object",
            "properties": {
                "token": {
                    "description": "JWT-токен для аутентификации",
                    "type": "string"
                }
            }
        },
        "v1.loginRequest": {
            "description": "Запрос для аутентификации пользователя",
            "type": "object",
            "properties": {
                "email": {
                    "description": "Электронная почта пользователя\nformat: email",
                    "type": "string"
                },
                "password": {
                    "description": "Пароль пользователя",
                    "type": "string"
                }
            }
        },
        "v1.loginResponse": {
            "description": "Ответ с токеном авторизации после успешной аутентификации",
            "type": "object",
            "properties": {
                "token": {
                    "description": "JWT-токен для аутентификации",
                    "type": "string"
                }
            }
        },
        "v1.registerRequest": {
            "description": "Запрос для регистрации нового пользователя",
            "type": "object",
            "properties": {
                "email": {
                    "description": "Электронная почта пользователя\nformat: email",
                    "type": "string"
                },
                "password": {
                    "description": "Пароль пользователя",
                    "type": "string"
                },
                "role": {
                    "description": "Роль пользователя (employee или moderator)\nenum: employee,moderator",
                    "type": "string"
                }
            }
        },
        "v1.registerResponse": {
            "description": "Ответ с данными зарегистрированного пользователя",
            "type": "object",
            "properties": {
                "email": {
                    "description": "Электронная почта пользователя\nformat: email",
                    "type": "string"
                },
                "id": {
                    "description": "Уникальный идентификатор пользователя\nformat: uuid",
                    "type": "string"
                },
                "role": {
                    "description": "Роль пользователя (employee или moderator)\nenum: employee,moderator",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}